Remove attendance through UI in OW4
