--- apps/events/templatetags/calendar_filters.py
@@ -1,10 +1,0 @@
-import re
-
-from django import template
-
-register = template.Library()
-
-
-@register.filter
-def unhttps(url):
-    return re.sub(r"^https", "http", url)
