+++ apps/events/views.py
@@ -2,252 +2,17 @@ import json
 
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.contrib.contenttypes.models import ContentType
-from django.core.signing import Signer
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
-from django.urls import reverse
-from django.utils import timezone
 from django.utils.translation import gettext as _
-from rest_framework import mixins, viewsets
-from rest_framework.permissions import AllowAny
-from watson import search as watson
 
-from apps.events.filters import EventFilter
-from apps.events.forms import CaptchaForm
-from apps.events.models import AttendanceEvent, Attendee, Event
+from apps.events.models import Event
 from apps.events.pdf_generator import EventPDF
-from apps.events.serializers import EventSerializer
 from apps.events.utils import (
-    handle_attend_event_payment,
-    handle_attendance_event_detail,
-    handle_event_ajax,
-    handle_event_payment,
     handle_mail_participants,
 )
-from apps.payment.models import Payment, PaymentDelay, PaymentRelation
 
 from .utils import EventCalendar
-
-
-def index(request):
-    context = {}
-    if request.user and request.user.is_authenticated:
-        signer = Signer()
-        context["signer_value"] = signer.sign(request.user.username)
-        context["personal_ics_path"] = request.build_absolute_uri(
-            reverse("events_personal_ics", args=(context["signer_value"],))
-        )
-    return render(request, "events/index.html", context)
-
-
-def details(request, event_id, event_slug):
-    event = get_object_or_404(Event, pk=event_id)
-
-    # Restricts access to the event if it is group restricted
-    if not event.can_display(request.user):
-        messages.error(request, "Du har ikke tilgang til dette arrangementet.")
-        return index(request)
-
-    if request.method == "POST":
-        if "action" in request.POST and "extras_id" in request.POST:
-            return JsonResponse(
-                handle_event_ajax(
-                    event,
-                    request.user,
-                    request.POST["action"],
-                    request.POST["extras_id"],
-                )
-            )
-
-    form = CaptchaForm(user=request.user)
-    context = {
-        "captcha_form": form,
-        "event": event,
-        "ics_path": request.build_absolute_uri(reverse("event_ics", args=(event.id,))),
-    }
-
-    if event.is_attendance_event():
-        try:
-            payment = Payment.objects.get(
-                content_type=ContentType.objects.get_for_model(AttendanceEvent),
-                object_id=event_id,
-            )
-        except Payment.DoesNotExist:
-            payment = None
-
-        context = handle_attendance_event_detail(event, request.user, context)
-        if payment:
-            request.session["payment_id"] = payment.id
-            context = handle_event_payment(event, request.user, payment, context)
-
-    return render(request, "events/details.html", context)
-
-
-def get_attendee(attendee_id):
-    return get_object_or_404(Attendee, pk=attendee_id)
-
-
-@login_required
-def attend_event(request, event_id):
-    event = get_object_or_404(Event, pk=event_id)
-
-    if not event.is_attendance_event():
-        messages.error(request, _("Dette er ikke et påmeldingsarrangement."))
-        return redirect(event)
-
-    if not request.POST:
-        messages.error(request, _("Vennligst fyll ut skjemaet."))
-        return redirect(event)
-
-    form = CaptchaForm(request.POST, user=request.user)
-
-    if not form.is_valid():
-        for errors in form.errors.values():
-            for error in errors:
-                messages.error(request, error)
-
-        return redirect(event)
-
-    # Check if the user is eligible to attend this event.
-    # If not, an error message will be present in the returned dict
-    attendance_event = event.attendance_event
-
-    response = event.attendance_event.is_eligible_for_signup(request.user)
-
-    if response.status:
-        attendee = Attendee(event=attendance_event, user=request.user)
-        if "note" in form.cleaned_data:
-            attendee.note = form.cleaned_data["note"]
-        attendee.show_as_attending_event = (
-            request.user.get_visible_as_attending_events()
-        )
-        attendee.save()
-        messages.success(request, _("Du er nå meldt på arrangementet."))
-
-        if attendance_event.payment():
-            handle_attend_event_payment(event, request.user)
-
-        return redirect(event)
-    else:
-        messages.error(request, response.message)
-        return redirect(event)
-
-
-@login_required
-def unattend_event(request, event_id):
-    event = get_object_or_404(Event, pk=event_id)
-
-    if not event.is_attendance_event():
-        messages.error(request, _("Dette er ikke et påmeldingsarrangement."))
-        return redirect(event)
-
-    attendance_event = event.attendance_event
-
-    # Check if user is attending
-    if len(Attendee.objects.filter(event=attendance_event, user=request.user)) == 0:
-        messages.error(request, _("Du er ikke påmeldt dette arrangementet."))
-        return redirect(event)
-
-    # Check if the deadline for unattending has passed
-    if (
-        attendance_event.unattend_deadline < timezone.now()
-        and not attendance_event.is_on_waitlist(request.user)
-    ):
-        messages.error(
-            request, _("Avmeldingsfristen for dette arrangementet har utløpt.")
-        )
-        return redirect(event)
-
-    if attendance_event.event.event_start < timezone.now():
-        messages.error(request, _("Dette arrangementet har allerede startet."))
-        return redirect(event)
-
-    try:
-        payment = Payment.objects.get(
-            content_type=ContentType.objects.get_for_model(AttendanceEvent),
-            object_id=event_id,
-        )
-    except Payment.DoesNotExist:
-        payment = None
-
-    # Delete payment delays connected to the user and event
-    if payment:
-        payments = PaymentRelation.objects.filter(
-            payment=payment, user=request.user, refunded=False
-        )
-
-        # Return if someone is trying to unatend without refunding
-        if payments:
-            messages.error(
-                request,
-                _(
-                    "Du har betalt for arrangementet og må refundere før du kan melde deg av"
-                ),
-            )
-            return redirect(event)
-
-        delays = PaymentDelay.objects.filter(payment=payment, user=request.user)
-        for delay in delays:
-            delay.delete()
-
-    Attendee.objects.get(event=attendance_event, user=request.user).delete()
-
-    messages.success(request, _("Du ble meldt av arrangementet."))
-    return redirect(event)
-
-
-def search_events(request):
-    query = request.GET.get("query")
-    filters = {
-        "future": request.GET.get("future"),
-        "myevents": request.GET.get("myevents"),
-    }
-    events = _search_indexed(request, query, filters)
-
-    return render(request, "events/search.html", {"events": events})
-
-
-def _search_indexed(request, query, filters):
-    results = []
-    kwargs = {}
-    order_by = "event_start"
-
-    if filters["future"] == "true":
-        kwargs["event_start__gte"] = timezone.now()
-    else:
-        # Reverse order when showing all events
-        order_by = "-" + order_by
-
-    if filters["myevents"] == "true":
-        kwargs["attendance_event__attendees__user"] = request.user
-
-    events = (
-        Event.objects.filter(**kwargs)
-        .order_by(order_by)
-        .prefetch_related(
-            "attendance_event",
-            "attendance_event__attendees",
-            "attendance_event__reserved_seats",
-            "attendance_event__reserved_seats__reservees",
-        )
-    )
-
-    # Filters events that are restricted
-    display_events = set()
-
-    for event in events:
-        if event.can_display(request.user):
-            display_events.add(event.pk)
-
-    events = events.filter(pk__in=display_events)
-
-    if query:
-        for result in watson.search(query, models=(events,)):
-            results.append(result.object)
-        return results[:10]
-
-    return events
 
 
 @login_required()
@@ -393,54 +158,3 @@ def mail_participants(request, event_id):
             "event": event,
         },
     )
-
-
-@login_required
-def toggle_show_as_attending(request, event_id):
-    event = get_object_or_404(Event, pk=event_id)
-
-    if not event.is_attendance_event():
-        messages.error(request, _("Dette er ikke et påmeldingsarrangement."))
-        return redirect(event)
-
-    attendance_event = event.attendance_event
-    attendee = Attendee.objects.get(event=attendance_event, user=request.user)
-
-    if attendee.show_as_attending_event:
-        attendee.show_as_attending_event = False
-        messages.success(
-            request, _("Du er ikke lenger synlig som påmeldt dette arrangementet.")
-        )
-    else:
-        attendee.show_as_attending_event = True
-        messages.success(request, _("Du er nå synlig som påmeldt dette arrangementet."))
-
-    attendee.save()
-    return redirect(event)
-
-
-class EventViewSet(
-    viewsets.GenericViewSet, mixins.RetrieveModelMixin, mixins.ListModelMixin
-):
-    serializer_class = EventSerializer
-    permission_classes = (AllowAny,)
-    filterset_class = EventFilter
-    filterset_fields = ("event_start", "event_end", "id")
-    ordering_fields = (
-        "event_start",
-        "event_end",
-        "id",
-        "has_passed",
-        "closest",
-    )
-    ordering = ("has_passed", "closest", "id")
-
-    def get_queryset(self):
-        user = self.request.user
-        return Event.by_nearest_active_event.get_queryset_for_user(user).select_related(
-            "image",
-            "organizer",
-            "group_restriction",
-            "attendance_event",
-            "attendance_event__reserved_seats",
-        )
