--- assets/events/archive/index.js
@@ -1,6 +1,0 @@
-import Filters from './filters';
-import './less/event_archive.less';
-
-const filters = new Filters();
-filters.bindEventListeners();
-filters.search();
