--- assets/events/archive/less/event_archive.less
@@ -1,103 +1,0 @@
-@import "assets/common/less/colors.less";
-
-#filter {
-  background: @baby-blue-medium;
-  padding: 10px;
-  color: #fff;
-  margin-bottom: 20px;
-  float: right;
-
-  input[type="text"] {
-    width: 100%;
-  }
-}
-
-#events {
-  #filter .input-group {
-    margin-bottom: 10px;
-    float: left;
-
-    input {
-      font-size: 12px;
-      border-radius: 15px;
-    }
-  }
-
-  .ingress {
-    margin-bottom: 10px;
-    line-height: 1.625em;
-  }
-
-  article {
-    margin-bottom: 40px;
-    padding-bottom: 30px;
-
-    /*border-bottom: 1px dotted #ccc;*/
-
-    /*position: relative;*/
-
-    button {
-      width: 100%;
-      margin: 2px 0 0 0;
-      border-radius: 0 0 0 0;
-    }
-
-    img {
-      width: 100%;
-    }
-
-    .content {
-      /*position: absolute;
-      top: 0;
-      left: 260px;*/
-    }
-
-    h1 {
-      font-size: 20px;
-      font-weight: 400;
-      color: #ee7810;
-      margin-top: 0;
-    }
-
-    span {
-      margin: 0 8px 0 0;
-      font-size: 18px;
-      text-transform: uppercase;
-      font-weight: 400;
-      color: #ee7810;
-      float: right;
-    }
-
-    .meta {
-      padding: 5px 0 0 0;
-      width: 100%;
-      border-top: 1px dotted #ccc;
-      font-size: 11px;
-      font-weight: normal;
-
-      /*text-align: left;*/
-
-      p {
-        color: #aaa;
-      }
-    }
-
-    @media (max-width: 989px) {
-      h1 {
-        font-size: 18px;
-      }
-
-      span {
-        float: left;
-        font-size: 14px;
-        margin-bottom: 10px;
-      }
-    }
-
-    @media (max-width: 767px) {
-      h1 {
-        margin-top: 15px;
-      }
-    }
-  }
-}
