--- assets/events/details/index.js
@@ -1,4 +1,0 @@
-import Events from './events';
-import './less/event_details.less';
-
-Events();
