+++ assets/frontpage/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 import ArticlesContainer from './articles/containers/ArticlesContainer';
-import EventsContainer from './events/containers/EventsContainer';
-import './initFrontpage';
 import './less/frontpage.less';
 
 const renderArticles = (Articles) => {
@@ -13,12 +11,3 @@ const renderArticles = (Articles) => {
 };
 
 renderArticles(ArticlesContainer);
-
-const renderEvents = (Events) => {
-  ReactDom.render(
-      <Events />,
-    document.getElementById('event-items'),
-  );
-};
-
-renderEvents(EventsContainer);
