+++ assets/frontpage/less/frontpage.less
@@ -4,4 +4,3 @@
 @import 'articles.less';
 @import 'business.less';
 @import 'company.less';
-@import 'events.less';
