+++ pyproject.toml
@@ -58,10 +58,10 @@ dev-dependencies = [
     "pytest-cov~=5.0",
     "pytest-django~=4.8",
     "pytest-xdist~=3.6",
-    "freezegun~=1.5",
     "coverage~=7.6",
     "ruff~=0.6",
     "pre-commit~=3.8",
+    "freezegun>=1.5.1",
 ]
 
 [tool.ruff]
