--- templates/events/extras.html
@@ -1,20 +1,0 @@
-{% if attendance_event.extras.all %}
-    <div class="dropdown extras">
-      <span class="extra-text"></span>
-      <select class="form-control extras-choice">
-        {% if not attendee.extras %}
-          <option value="" disabled selected>Velg et alternativ</option>
-        {% endif %}
-        {% for choice in attendance_event.extras.all %}
-            <option value="{{ choice.id }}"
-              {%if attendee.extras == choice %}
-                selected="selected">
-                Valgt ekstra: {{choice}}
-              {%else%}
-                >{{ choice }}
-              {%endif%}
-            </option>
-        {% endfor %}
-      </select>
-    </div>
-{% endif %}
