--- templates/events/index.html
@@ -1,103 +1,0 @@
-{% extends "base.html" %}
-
-{% load calendar_filters %}
-{% load render_bundle from webpack_loader %}
-
-{% block title %}
-Arrangementsarkiv - Online
-{% endblock title %}
-
-{% block styles %}
-  {{ block.super }}
-  {% render_bundle 'eventsArchive' 'css' %}
-{% endblock %}
-
-{% block submenu %}
-    <nav class="subnavbar">
-        <div class="container">
-            <ul>
-                <li class="top-menu-link active">
-                    <a href="{% url 'events_index' %}">Arrangementer</a>
-                </li>
-                <li class="top-menu-link">
-                    <a href="{% url 'article_archive' %}">Artikler</a>
-                </li>
-            </ul>
-        </div>
-    </nav>
-{% endblock %}
-
-{% block content %}
-
-<section>
-  <div class="container">
-
-    <div class="row">
-      <div class="col-md-12">
-        <div class="page-header">
-          <h1>Arrangementer
-            <div class="btn-group pull-right">
-              <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
-                <span class="glyphicon glyphicon-calendar"></span>
-                Eksporter
-                <span class="caret"></span>
-              </button>
-              <ul class="dropdown-menu">
-                <li><a href="{% url 'events_ics' %}">ICS</a></li>
-                <li><a href="http://www.google.com/calendar/render?cid={{ request.build_absolute_uri|unhttps|urlencode }}events.ics">Google Calendar</a></li>
-                {% if user.is_authenticated %}
-                  <li role="presentation" class="divider"></li>
-                  <li class="dropdown-header">Mine arrangementer</li>
-                  <li><a href="{% url 'events_personal_ics' user=signer_value %}">ICS</a></li>
-                  <li><a href="http://www.google.com/calendar/render?cid={{ personal_ics_path|unhttps|urlencode }}">Google Calendar</a></li>
-                  {% endif %}
-              </ul>
-            </div>
-          </h1>
-        </div>
-      </div>
-    </div>
-
-    <div class="col-xs-12 col-sm-12 col-md-3 pull-right" id="filter">
-      <div class="row">
-        <div class="col-xs-12 col-sm-12 col-md-12">
-          <input type="text" class="form-control col-xs-12 col-sm-12" placeholder="Søk blant arrangementene" id="search">
-        </div>
-      </div>
-
-
-        <div class="checkbox col-xs-12 col-sm-12 col-md-12">
-          <label>
-            <input type="checkbox" name="optionCheckboxs" value="option1" id="future" checked>
-            Bare fremtidige arrangementer
-          </label>
-        </div>
-	{% if user.is_authenticated %}
-        <div class="checkbox col-xs-12 col-sm-12 col-md-12">
-          <label>
-            <input type="checkbox" name="optionCheckboxs" value="option1" id="myevents">
-            Mine arrangementer
-          </label>
-        </div>
-	{% endif %}
-
-    </div> <!-- end col-md-3 -->
-
-    <div class="col-xs-12 col-sm-12 col-md-9 pull-left" id="events">
-
-    </div> <!-- end col-md-9 -->
-<!--
-  {% if events %}
-
-  {% else %}
-    <p class="ingress">Ingen arrangementer funnet.</p>
-  {% endif %} -->
-
-  </div>
-</section>
-{%endblock content %}
-
-{% block js %}
-  {{ block.super }}
-  {% render_bundle 'eventsArchive' 'js' attrs='async type="module"' %}
-{% endblock %}
