--- templates/events/payment_tag.html
@@ -1,17 +1,0 @@
-<a
-    href="https://online.ntnu.no/payments/events/{{ payment.object_id }}"
-    class="btn btn-large btn-success payment-button"
-    target="_blank"
-    rel="noopener noreferrer"
->
-    Gå til betalingssiden
-</a>
-<br />
-
-{% if payment_delay %}
-<p class="help-text row_space">Betalingsfrist: {{ payment_delay.valid_to }}</p>
-<br />
-{% elif payment.deadline %}
-<p class="help-text row_space">Betalingsfrist: {{ payment.deadline }}</p>
-<br />
-{% endif %}
