--- templates/events/search.html
@@ -1,78 +1,0 @@
-{% load markdown_deux_tags %}
-{% load shuffle %}
-
-{% if events %}
-  {% for event in events %}
-  <article data-date="{{ event.event_start|date:"c" }}" class="row">
-
-    <div class="col-xs-12 col-sm-4 col-md-4 left-box">
-      <div class="row">
-        {% if event.companies.all %}
-            {% with event.companies.all|shuffle|first as random_relation %}
-                <a href="{{ event.get_absolute_url }}">
-                  <picture>
-                    <source srcset="{{ random_relation.company.image.lg }}" media="(max-width: 768px)">
-                    <img src="{{ random_relation.company.image.md }}" alt="" />
-                  </picture>
-                </a>
-            {% endwith %}
-        {% else %}
-            <a href="{{ event.get_absolute_url }}">
-              <picture>
-                <source srcset="{{ event.image.lg }}" media="(max-width: 768px)">
-                <img src="{{ event.image.md }}" alt="" />
-              </picture>
-            </a>
-        {% endif %}
-        {% if event.is_attendance_event %}
-            <a href="{{ event.get_absolute_url }}">
-                <button type="button" class="btn btn-success">Gå til påmelding</button>
-            </a>
-        {% endif %}
-      </div>
-    </div>
-
-    <div class="col-xs-12 col-sm-8 col-md-8 content">
-        <div class="row">
-            <div class=" col-sm-12 col-md-8">
-                <h1>
-                    <a href="{{ event.get_absolute_url }}">{{ event.title }}</a>
-                </h1>
-            </div>
-            <div class="col-sm-12 col-md-4">
-                <span>{{ event.event_start|date:"d.m.Y"}}</span>
-            </div>
-        </div>
-
-      <div class="ingress">{{ event.ingress|striptags|markdown }}</div>
-
-      <div class="meta">
-        {% if event.is_attendance_event %}
-          {% if event.attendance_event.number_of_seats_taken == event.attendance_event.max_capacity %}
-		  <div class="col-md-3">
-		    <p>Venteliste: {{ event.attendance_event.number_on_waitlist }} </p>
-		  </div>
-		  {% else %}
-          <div class="col-md-3">
-            <p>Påmeldte: {{ event.attendance_event.number_of_seats_taken }} / {{ event.attendance_event.max_capacity }}</p>
-          </div>
-          {% endif %}
-          <div class="col-md-3">
-            <p>Sted: {{ event.location }}</p>
-          </div>
-          <div class="col-md-5">
-            <p>Påmelding åpner {{ event.attendance_event.registration_start|date:"d.m.y H.i" }}</p>
-          </div>
-        {% else %}
-          <div class="col-md-12">
-            <p>Sted: {{ event.location }}</p>
-          </div>
-        {% endif %}
-      </div>
-    </div>
-
-  </article>
-  {% endfor %}
-{% else %}
-  <p class="ingress">Ingen arrangementer funnet, eller de er blokkert av en adblock.</p>
-{% endif %}
